+++ framework/db-http/src/main/java/is/codion/framework/db/http/HttpEntityConnection.java
@@ -24,7 +24,6 @@ import is.codion.framework.db.EntityConnection;
 import is.codion.framework.domain.DomainType;
 
 import java.util.UUID;
-import java.util.concurrent.Executor;
 
 import static is.codion.common.Configuration.*;
 
@@ -199,13 +198,6 @@ public interface HttpEntityConnection extends EntityConnection {
 		 * @return this builder instance
 		 */
 		Builder clientId(UUID clientId);
-
-		/**
-		 * By default, the http client uses a shared thread pool executor.
-		 * @param executor the http client executor to use
-		 * @return this builder instance
-		 */
-		Builder executor(Executor executor);
 
 		/**
 		 * @return a http based EntityConnection